requirements.txt
Pillow>=10.0.0
mido>=1.3.0
numpy>=1.24.0
pytest>=7.0.0
